Added Exceptions to the Project-Service Controller add, delete, findAll

+++ src/main/java/com/revature/controllers/ProjectController.java
@@ -1,12 +1,12 @@
 package com.revature.controllers;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,7 +15,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.revature.exceptions.ProjectNotAddedException;
+import com.revature.exceptions.ProjectNotFoundException;
 import com.revature.models.Project;
+import com.revature.models.ProjectErrorResponse;
 import com.revature.service.ProjectService;
 
 @RestController
@@ -61,13 +64,23 @@ public class ProjectController {
 	@PostMapping("/add")
 	@ResponseStatus(HttpStatus.OK)
 	public Boolean addProject(@RequestBody Project project) {
-		return 	projectService.addProject(project);
+		if (project.getBatch() == null)
+			throw new ProjectNotAddedException("The 'batch' input cannot be null when adding project");
+		if (project.getName() == null)
+			throw new ProjectNotAddedException("The 'name' input cannot be null when adding project");
+		if (project.getTechStack() == null)
+			throw new ProjectNotAddedException("The 'tech stack' input cannot be null when adding project");
+		return projectService.addProject(project);
 	}
 
 	// Delete by ID
 	@DeleteMapping("/delete/{id}")
 	@ResponseStatus(HttpStatus.OK)
 	public Boolean deleteById(@PathVariable String id) {
+		Project ID = projectService.findById(id);
+		if (ID == null) {
+			throw new ProjectNotFoundException("ID entered cannot be found to delete this project");
+		}
 		return projectService.deleteById(id);
 	}
 
@@ -77,12 +90,36 @@ public class ProjectController {
 	public Boolean updateProject(@RequestBody Project project) {
 		return projectService.updateProject(project);
 	}
-	
-	// Update Project
-		@GetMapping("/{id}")
-		@ResponseStatus(HttpStatus.OK)
-		public Project updateProject(@PathVariable String id) {
+
+	// Find By ID
+	@GetMapping("/{id}")
+	@ResponseStatus(HttpStatus.OK)
+	public Project updateProject(@PathVariable String id) {
+		Project ID = projectService.findById(id);
+			if (ID == null) {
+				throw new ProjectNotFoundException("ID entered cannot be found");
+			}
 			return projectService.findById(id);
-			
 		}
+	
+
+	@ExceptionHandler
+	@ResponseStatus(HttpStatus.NOT_FOUND)
+	public ProjectErrorResponse handleExceptions(ProjectNotFoundException pnfe) {
+		ProjectErrorResponse error = new ProjectErrorResponse();
+		error.setStatus(HttpStatus.NOT_FOUND.value());
+		error.setMessage(pnfe.getMessage());
+		error.setTimmeStamp(System.currentTimeMillis());
+		return error;
+	}
+
+	@ExceptionHandler
+	@ResponseStatus(HttpStatus.BAD_REQUEST)
+	public ProjectErrorResponse handleExceptions(ProjectNotAddedException pnae) {
+		ProjectErrorResponse error = new ProjectErrorResponse();
+		error.setStatus(HttpStatus.BAD_REQUEST.value());
+		error.setMessage(pnae.getMessage());
+		error.setTimmeStamp(System.currentTimeMillis());
+		return error;
+	}
 }
